Add stubs fixtures and refactor tests
